Add constructor for GameViewResolution attribute
Added a new constructor to the GameViewResolutionAttribute class that takes a GameViewResolution enum as a parameter instead of width, height and a name. This simplifies the application of this attribute in the GameViewResolutionAttributeTest as you can directly pass the predefined enum value. This change aims to improve code readability and ease of use.