Update Unity version in test workflow
The Unity versions used in the .github workflow for testing were updated. Earlier, the versions were 2022.3.4f1 and 2023.1.2f1. They have been changed to 2022.3.11f1 and 2023.1.16f1, respectively. This change is necessary to ensure that the tests are run using the recent Unity versions, thus keeping our package up-to-date and relevant.